Still needing to fix setCurrentUser type

+++ src/components/SignOut.tsx
@@ -1,16 +1,24 @@
-import React, { FC } from 'react';
+import React, { FC, useContext } from 'react';
+import { AtucasaContext } from '../Context';
 
 const SignOut: FC = () => {
-  
+  const { setCurrentUser, currentUser } = useContext<any>(AtucasaContext);
+
+
   const handleSignOut = (event:React.MouseEvent<HTMLButtonElement>):void => {
     event.preventDefault();
-    fetch("http://localhost:3000/logout", {
-      credentials: "include",
-      method: "DELETE"
-    })
-    .then(response => response.json())
-    .then(console.log)
-    .catch(err => console.error(err))
+    if (currentUser) {
+      fetch("http://localhost:3000/logout", {
+        credentials: "include",
+        method: "DELETE"
+      })
+      .then(response => response.json())
+      .then(data => {
+        console.log(data);
+        setCurrentUser(null);
+      })
+      .catch(err => console.error(err))
+    } 
   }
 
   return (
